29-Mar-2017 email.  Details in description
29-Mar-2017 email 1/2:
Hi Robert
Last year you kindly put the GP earnings on the website
Could you change this to the following:
This needs to be done by the 31st March if that is possible.  Could you
confirm when it is done.  Thanks.
Lesley
29-Mar-2017 email 2/2
Thanks – could you also check to see if it says whether prescriptions
are ready for collection in 2 days if so could you change it to 3 for
me.  Thanks you are a star.